Upgrade to Hibernate 3.6 / Release Notes 3.3->3.5: change from int to long

+++ src/main/java/org/linagora/linshare/core/business/service/DocumentEntryBusinessService.java
@@ -83,7 +83,7 @@ public interface DocumentEntryBusinessService {
 
 	public List<ThreadEntry> findAllThreadEntriesTaggedWith(Thread owner, String[] names);
 
-	public int countThreadEntries(Thread thread);
+	public long countThreadEntries(Thread thread);
 
 	public InputStream getDocumentStream(ThreadEntry entry);
