+++ src/main/java/org/linagora/linshare/core/facade/impl/ThreadEntryFacadeImpl.java
@@ -211,7 +211,7 @@ public class ThreadEntryFacadeImpl extends GenericTapestryFacade implements Thre
 	}
 
 	@Override
-	public int countEntries(ThreadVo threadVo) throws BusinessException {
+	public long countEntries(ThreadVo threadVo) throws BusinessException {
 		return threadService.countEntries(findThread(threadVo));
 	}
 
@@ -352,7 +352,7 @@ public class ThreadEntryFacadeImpl extends GenericTapestryFacade implements Thre
 	}
 
 	@Override
-	public int countMembers(UserVo actorVo, ThreadVo threadVo)
+	public long countMembers(UserVo actorVo, ThreadVo threadVo)
 			throws BusinessException {
 		return threadService.countMembers(findThread(threadVo));
 	}
